🤖 Move activity test to avoid including Fleet core activities from the activites module
Zed + Opus 4.6; prompt: Move `TestVPPInstallFailureEmptyCommandUUIDDoesNotActivateNext` to the main module to avoid the activity module depending on the main module.
